Vaccination: add edit command
